refresh offline activitiy layer on changes
